feat(dropdown): added new prop "movePopupAsYouType"; default caretPosition is "next" from now
This commint change the default behavior of the RTA. From this very moment, the textarea will stay
on the place when it occurs for the first time. You can enforce the old behavior with a new prop
"movePopupAsYouType". This change has also affected the default styles a bit, co be aware of it.
I've also changed default setting for "caretPosition" to "next" value from old one "end".

BREAKING CHANGE popup will not follow caret as it used to; default "caretPosition" is "next"